+++ frontend/src/App.jsx
@@ -1,6 +1,6 @@
 import './App.css';
 import { BrowserRouter as Router, Routes, Route, useNavigate } from 'react-router-dom';
-import ForgottenThings from './components/ForgottenThings';
+import ForgottenThings from './pages/ForgottenThings';
 
 function Home() {
   const navigate = useNavigate();
